added functionality to create, delete, and edit user

+++ user/user_list.php
@@ -26,29 +26,32 @@
             if(isset($list_of_users) && is_array($list_of_users)){
                 foreach($list_of_users as $user){
                     $fa_cls = ($user->get_status()==1?'fa-go':'fa-stop');
-                    echo '<tr>
-                    <td><i class="fa fa-certificate '.$fa_cls.'" aria-hidden="true"></i></td>
-                    <td>'.$user->get_firstname().'</td>
-                    <td>'.$user->get_lastname().'</td>
-                    <td>'.$user->get_type().'</td>
-                    <td>'.$user->get_username().'</td>
-                    <td>'.$user->get_status_def().'</td>
-                    <td>
-                        <a href="functions/admin.php?v=edit&id='.$user->get_person_id().'">
-                            <i class="fa fa-pencil" aria-hidden="true" title="Edit"></i>
-                        </a>
-                    </td>
-                    <td>
-                        <a href="functions/admin.php?v=delete&id='.$user->get_person_id().'">
-                            <i class="fa fa-trash" aria-hidden="true" title="Delete"></i>
-                        </a>
-                    </td>
-                    <td>
-                        <a href="functions/admin.php?v=suspend&id='.$user->get_person_id().'&cs='.$user->get_status().'">
-                            <i class="fa fa-pause" aria-hidden="true" title="Suspend"></i>
-                        </a>
-                    </td>
-                </tr>';
+
+                    if($user->get_username() != $_SESSION['user']){
+                        echo '<tr>
+                        <td><i class="fa fa-certificate '.$fa_cls.'" aria-hidden="true"></i></td>
+                        <td>'.$user->get_firstname().'</td>
+                        <td>'.$user->get_lastname().'</td>
+                        <td>'.$user->get_type().'</td>
+                        <td>'.$user->get_username().'</td>
+                        <td>'.$user->get_status_def().'</td>
+                        <td>
+                            <a href="'.$_SERVER['PHP_SELF'].'?v=edit&id='.$user->get_person_id().'">
+                                <i class="fa fa-pencil" aria-hidden="true" title="Edit"></i>
+                            </a>
+                        </td>
+                        <td>
+                            <a href="functions/admin.php?v=delete&id='.$user->get_person_id().'">
+                                <i class="fa fa-trash" aria-hidden="true" title="Delete"></i>
+                            </a>
+                        </td>
+                        <td>
+                            <a href="functions/admin.php?v=suspend&id='.$user->get_person_id().'&cs='.$user->get_status().'">
+                                <i class="fa fa-pause" aria-hidden="true" title="Suspend"></i>
+                            </a>
+                        </td>
+                    </tr>';
+                    }
                 }
             }
